# src/docs/openapi.yaml
...
openapi: 3.0.0
info:
  title: Marketplace API Documentation (Final)
  version: 1.0.0
  description: Dokumentasi API untuk sistem Kalana Commerce, mencakup Auth, Produk, Cart, dan Multi-Order.
servers:
  - url: http://202.155.14.40/api
    description: Production
  - url: http://localhost:3000/api
    description: Development Server (Lokal)
  

# ==========================================================
# DEFINISI JALUR (PATHS)
# ==========================================================
paths:
...
  /me:
    get:
      tags:
        - 1. User Authentication
      summary: Mendapatkan Profil User yang sedang Login (Requires JWT)
      security:
        - bearerAuth: [] # ⬅️ Butuh Auth JWT
      responses:
        '200':
          description: Profil user berhasil diambil, termasuk alamat default.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          description: Unauthorized (Token missing/invalid)
  /addresses:
    get:
      tags:
        - User Account & Addresses
      summary: Mendapatkan SEMUA Alamat Milik User yang Login
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Daftar semua alamat user berhasil diambil.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AddressDetail'

    post:
      tags:
        - 5. User Account & Addresses
      summary: Menambahkan Alamat Baru
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddressInput'
      responses:
        '201':
          description: Alamat berhasil dibuat.
        '422':
          description: Validasi Gagal.
...
  /carts:
    get:
      tags:
        - 3. Cart Management
      summary: Mendapatkan Semua Item Cart User
      security:
        - bearerAuth: [] # ⬅️ Butuh Auth JWT
      responses:
        '200':
          description: Daftar item keranjang (termasuk detail produk dan outlet) berhasil diambil.
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Masukkan token JWT dalam format "Bearer [token]"
  schemas:
...
    UserResponse:
      type: object
      description: Skema data user yang dikembalikan oleh endpoint /me.
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        email:
          type: string
          format: email
        phoneNumber:
          type: string
        balance:
          type: number
          format: float
        createdAt:
          type: string
          format: date-time
        # Alamat Default yang disertakan (max 1 item)
        addresses:
          type: array
          items:
            $ref: '#/components/schemas/AddressDetail'
          description: Alamat default pengguna (hanya 1 item jika ada).
    AddressDetail:
      type: object
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        label: { type: string, example: "Rumah Utama" }
        recipientName: { type: string, example: "Nurhayati" }
        phoneNumber: { type: string, example: "08901234567" }
        street: { type: string, example: "Jl. Raya Mawar No. 5" }
        postalCode: { type: string, example: "55123" }
        provincesId: { type: string, example: "ID-JI" }
        citiesId: { type: string, example: "ID-JI-SBY" }
        isDefault: { type: boolean, example: true }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        deletedAt: { type: string, format: date-time, nullable: true }
    AddressInput:
      type: object
      required: [label, recipientName, phoneNumber, street, postalCode, provincesId, citiesId, districtsId]
      properties:
        label: { type: string, example: "Kantor Cabang" }
        recipientName: { type: string, example: "Satria Tama" }
        phoneNumber: { type: string, example: "08901234567" }
        street: { type: string, example: "Gedung Sudirman Lt. 10" }
        postalCode: { type: string, example: "10220" }
        provincesId: { type: string, example: "ID-JK" }
        citiesId: { type: string, example: "ID-JK-JKT" }
        districtsId: { type: string, example: "ID-JK-JKTS" }
        isDefault: { type: boolean, example: false }
        long: { type: number, format: float }
        lat: { type: number, format: float }
...